Snackbar message module
