+++ products/jbrowse-desktop/src/sessionModelFactory.ts
@@ -7,10 +7,10 @@ import {
 } from '@jbrowse/core/configuration'
 import {
   Region,
-  NotificationLevel,
   TrackViewModel,
   DialogComponentType,
 } from '@jbrowse/core/util/types'
+import addSnackbarToModel from '@jbrowse/core/ui/SnackbarModel'
 import { getContainingView } from '@jbrowse/core/util'
 import { observable } from 'mobx'
 import {
@@ -545,30 +545,7 @@ export default function sessionModelFactory(
         return getParent(self).setSession(sessionSnapshot)
       },
     }))
-    .extend(() => {
-      const snackbarMessages = observable.array()
 
-      return {
-        views: {
-          get snackbarMessages() {
-            return snackbarMessages
-          },
-        },
-        actions: {
-          notify(message: string, level?: NotificationLevel) {
-            return this.pushSnackbarMessage(message, level)
-          },
-
-          pushSnackbarMessage(message: string, level?: NotificationLevel) {
-            return snackbarMessages.push([message, level])
-          },
-
-          popSnackbarMessage() {
-            return snackbarMessages.pop()
-          },
-        },
-      }
-    })
     .views(self => ({
       getTrackActionMenuItems(config: any) {
         const session = self
@@ -620,7 +597,7 @@ export default function sessionModelFactory(
       },
     }))
 
-  return types.snapshotProcessor(sessionModel, {
+  return types.snapshotProcessor(addSnackbarToModel(sessionModel), {
     // @ts-ignore
     preProcessor(snapshot) {
       if (snapshot) {
